daemon: set titles for subprocesses
For every new process, replace its `argv` array with a single string
clarifying its role.  This string will be visible in `ps` output and
similar.  Also, log a LOG_DEBUG message so the role of every process is
obvious from the syslog contents.

The job executor processes use the job type name instead of the role.