Pass chroot directory as a file descriptor
Since the architecture of hasher-priv changed from a set-UID executable
to a client-server, there is a chance that the client could be run in
a chrooted environment.  Consequently, the client view of the chroot
directory can differ from the server one.  Add support for this scenario
by passing the chroot directory as a file descriptor instead of
a pathname.

This change is a part of nested chroots support.