chrootuid: initialize groups for satellite users
If the target user is also a caller user in the inner chroot, it would
need its supplementary groups set to initialize and access the inner
chroot properly.

Before this change, chrootuid was just clearing the set of supplementary
groups before chroot.

After this change, the set of supplementary groups for the target user
is calculated before entering chroot and assigned after entering chroot
right before the setgid() invocation.

This change is a part of nested chroots support.