daemon: use uids in the caller config file names
Since the architecture of hasher-priv changed from a set-UID executable
to a client-server, there is a chance that the client could be run in
a chrooted environment.  Consequently, the client view of the passwd
database can differ from the server one.  Add support for this scenario
by switching to use uids rather than usernames in the caller config file
names, but retain support of usernames there for backwards
compatibility.

This change is a part of nested chroots support.