job: put all request handlers in the cgroup of the client
Since the architecture of hasher-priv changed from a set-UID executable
to a client-server, it's likely that the client and the server belong
to different cgroups.  Avoid potential escape from the client's cgroup
by changing the server to join the client's cgroup.

If the job is a chrootuid, it will spawn unprivileged processes that can
take as much resources as they are allowed to.  Therefore, chrootuid
jobs are moved to the cgroup of the client so the cgroup regulations
of the client apply to these processes.  Other kinds of jobs are
short-lived processes that perform very specific auxiliary tasks,
they are parts of the service daemon and remain in its cgroup.

Note that cgroup1 API is not supported in favour of cgroup2.

Co-authored-by: Dmitry V. Levin <ldv@altlinux.org>