chrootuid: check namespaces
Since the architecture of hasher-priv changed from a set-UID executable
to a client-server, there is a chance that the client and the server
belong to different namespaces.  Avoid potential escape from the client's
namespaces by checking that both the client and the server belong to the
same namespaces.

Subsequent commits would implement support for some namespace
differences between the client and the server.