chrootuid: enter the client's mount, ipc, uts, and network namespaces
Since the architecture of hasher-priv changed from a set-UID executable
to a client-server, there is a chance that the client and the server
belong to different namespaces, for example, the client could be run
with its mount, ipc, uts and network namespaces isolated by another
hasher.  Add support for this scenario by changing the server to join
the client's mount, ipc, uts, and network namespaces if they differ.

This change is a part of nested chroots support.