spawn_job_runner: wait a bit for spawning jobs
We cannot wait for chrootuid jobs like we wait for other kinds of jobs
because chrootuid jobs can run for a long time controlled by the caller.

But we have to wait for chrootuid jobs at least until they joined the
cgroup of the client to make sure we do not attempt to start more
chrootuid jobs than the cgroup of the client allows to.

Actually we will wait until the executor has sanitized its descriptors.