Package SOCKETDIR
Package both the socket directory itself and the tmpfiles rules
required to re-create it at OS startup.

This change is not required for the daemon because its startup scripts
already make sure the directory is created.  The only case when this
change might help is when the socket directory is used before the daemon
is started, e.g. in a nested chroot configuration.