hasher-useradd: enable hasher-privd if users were added
Some users are complaining that since 2.0 there is an additional step in
setting up hasher-priv for the first time: the privileged service has to
be manually started/enabled by the administrator.

One way to address this is to make the service socket-activatable, but
currently in Sisyphus this is a systemd-only approach. (I am not sure if
xinetd in the repository can listen on unix sockets, or if the entire
sysvrc-xinetd setup is not too unwieldy)

This patch makes hasher-useradd enable and start the service if the
program has successfully added the 2 auxiliary users.

Reported-by: George Kouryachy <george@altlinux.org>
Suggested-by: Dmitry V. Levin <ldv@altlinux.org>