Add the experimental nproc system setting
This is intended for shared servers with a high CPU count, in order to
limit the per-hasher instance nproc to a reasonable value.

This function operates by randomizing the CPU affinity mask to match the
specified nproc count. Thanks to this randomization process, users'
tasks will be evenly distributed across all available CPUs.

Signed-off-by: Vitaly Chikunov <vt@altlinux.org>